SU-6986: Decrease service WMI query execution time
Users have reported significant execution times when running the task
and this script seems to be the culprit. By only selecting relevant
services, and only the properties we need, the query takes 1/3 of the
original time even on a small box with plenty of resources.

CIM was investigated, but the performance increase over the old Cmdlet
was non-noticable and not worth the loss of PSv2 support for the script
main execution path.